+++ cmd/helper/mingw-build.ts
@@ -393,6 +393,7 @@ const sumatraFiles: FileGroup[] = [
       "RegistryPreview.*",
       "RegistrySearchFilter.*",
       "SearchAndDDE.*",
+      "ExplorerQuickLook.*",
       "Screenshot.*",
       "ScreenshotCapture.*",
       "Selection.*",